refactor: Move policy report generation to backend
- Extract report generation logic from the frontend into a new backend
utility (policyReportGenerator.js).
- The POST /generateReport route now
builds and returns the HTML file directly, with the filename set via
Content-Disposition.
- The frontend utility is reduced to a thin fetch call that triggers the browser download from the response blob.
